+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Version 0.7.6
 
+- Added exception handling to analysis functions
+
 - Updated saveLoadV1 examples to use compactConnFormat=True and saveCellSecs=False 
 
 - Added cfg.rand123GlobalIndex to set global index used by all instances of the Random123 instances of Random